generation_scripts/requirements-demo.txt
reflex>=0.5.0
reflex-clerk>=0.1.0
sqlmodel>=0.0.8
alembic>=1.13.0
python-dotenv>=1.0.0
pyodbc>=5.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0
httpx>=0.25.0
